Major reorg.
The inbuilt debugger thing was a neat idea, but it wasn't working out as
I expected.  Moving the entire codebase so it is a library/tool/app or
framework/tool/app combination instead.

This means a total re-design of most of the code, as well as meaning
that Number and String have to integrate with NSNumber and NSString.

This is going to be fun.